fix queue lengths for non gate queues
After the introduction of the Depends-On visualization in zuul, dash
has been over counting the length of check and other queues by
counting the dependent patches that don't have tests being run on
them. This only counts the head length on the gate queue, and clips
the rest of them to only the last element in a head.