fix for gate pipeline display
pipelines are attributes of changes, but it's possible for new changes
added to a queue, the pipeline is not yet added (especially if it's
outside of the sliding window for that gate).

This looks at the pipeline for the first job of every change in the
queue, and ensures is true if 'gate' is the pipeline for any of them.