Update the readme about using --passwd for http auth
dash uses http auth since 9e2ad586ddc6826d0e9a60a292e746a71c3691ec.

For people like me that didn't realize this and expected it to just
keep working with my public ssh key that I use for gerrit, it's
confusing when you get a 401.

So this updates the README to point out that you need to have a
valid http password in your gerrit settings and use that with the -P
option now to connect.