Add the ability to specify multiple changes
Sometimes it's nice to be able to monitor multiple
changes from different owners in different projects.
This changes the -c option to be a list so you can
specify multiple changes at once.

Since the gerrit query language only handles multiple
changes with an OR (AND means none are shown), if multiple
changes are shown, the operator option is hard-coded to 'OR'.